+++ src/main/java/ch/ipt/JpaSwaggerLearning/service/UserService.java
@@ -1,8 +1,8 @@
 package ch.ipt.JpaSwaggerLearning.service;
 
 import ch.ipt.JpaSwaggerLearning.model.UserEntity;
-import ch.ipt.JpaSwaggerLearning.openapi.user.model.UserCreateDTO;
-import ch.ipt.JpaSwaggerLearning.openapi.user.model.UserDTO;
+import ch.ipt.JpaSwaggerLearning.openapi.model.UserCreateDTO;
+import ch.ipt.JpaSwaggerLearning.openapi.model.UserDTO;
 import ch.ipt.JpaSwaggerLearning.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -33,7 +33,7 @@ public class UserService {
     public UserDTO getUserById(int id){
         UserDTO userDTO = new UserDTO();
         UserEntity entity = userRepository.getReferenceById(id);
-        userDTO.setId(entity.getId().longValue());
+        userDTO.setId(entity.getId());
         userDTO.setName(entity.getName());
 
         return userDTO;
